fix(studioPage): 폴더 구조 재정의
- /studio/[user] -> /studio/[uid] 로 동적 경로 폴더이름 변경
- 대시보드 UI 배치 및 라우트 정의
- 대시보드 UI 구성에 따른 (route)/studio 폴더 구조화

Resolves: #27
