+++ src/app/(route)/studio/[uid]/_components/Header/Header.server.tsx
