+++ src/app/(route)/studio/[uid]/_components/Navigation/Navigation.client.tsx
