+++ src/app/(route)/studio/[uid]/_components/common/StudioWrapper.server.tsx
