--- src/app/(route)/studio/[user]/page.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const StudioPage = () => {
-  return <div>content</div>;
-};
-
-export default StudioPage;
-
-
